requirements.txt
pytorch-lightning==1.5.7
# torch==1.9.0+cu111
# torchvision==0.10.0+cu111
hydra-core==1.1.1
python-dotenv~=0.15.0
wandb==0.12.9
pandas==1.2.3
scikit-learn==0.24.1
rich==10.16.1
numpy==1.20.3
matplotlib==3.5.1
scikit-plot==0.3.7
seaborn==0.11.2
torchmetrics==0.6.2


black==21.12b0
isort==5.10.1
mypy==0.930
flake8==4.0.1


